pyproject.toml
[tool.poetry]
name = "fastoblig"
version = "0.1.0"
description = ""
authors = ["Patrick Stünkel <webminz@googlemail.com>"]
readme = "README.md"

[tool.poetry.scripts]
fastoblig = "fastoblig.main:app"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
openai = "^1.14.2"
python-dotenv = "^1.0.1"
pygit2 = "^1.14.1"
gitpython = "^3.1.43"
pytest = "^8.1.1"
typer = "^0.15.2"
polars = "^1.24.0"


[tool.poetry.group.dev.dependencies]
types-pygit2 = "^1.14.0.20240317"
types-requests = "^2.31.0.20240311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

